Don't store blobs in the cached state
This introduces state version 2. If a version 1 state is encountered, it
will be pruned of all blob IDs.

The immediate result of this commit is inefficient because it causes
extra file content lookups against the hg repo. However, future patches
will rewrite that code and the end result will be more efficient.